Changing ConcurrentBufferedUpdate to accommodate the new neuron group recorder issue.
Concurrent buffered update has been changed so it will work with the
new recording framework. It now also automatically folds in input
groups that read from files. Apply current in SpikingNeuronUpdateRule
was replaced with the already extant, inputValue variable in Neuron
which made it redundant. ConcurrentBufferedUpdate was also cleaned up
substantially as was the concurrency tools package to remove unused or
useless methods/classes. The cyclic barrier in ConcurrentBufferedUpdate
was replaced with a more efficient and lighter weight version that
ensures synchronization and is dependent on the new
SynchronizationPoint class which has been custom built for Simbrain’s
particular synchronizing needs.