Fixing Error with loading/saving synapse groups also STDP
The error with loading/saving synapse groups has been fixed.
Furthermore it is a solution that also addresses the OutOfMemeory
exception that can occur for large groups as this solution does not
require the construction of a full uncompressed weight matrix prior to
compression.

Also tossing in a fix to STDP. STDP now only changes weight in response
to a pair of spikes once, instead of forever. It also doesn’t do any of
the checks it used to as it instead simply checks if the source or
target have spiked. Now there’s just a try/catch if someone tries to
use a non-spiking neuron. This does not affect performance when
exceptions are not thrown.