Editing sparse connection panel to get rid of bug
Get rid of bug where an IndexOutOfBoundsException would pop up when
opening the network panel. The issue was occurred when the sparse
dialog was used on a sparse connection where no source or target
neurons were specified, which was almost always the case for setting
the sparse connector which would be used for quick connects (which is
never associated with any particular sources/targets).