More cleanups associated with RealNNAdEx script
Added methods in network that assign labels to network, also a helper
method that returns all neuron and synapse groups (including those
inside subnetworks), but not subnetworks themselves, which is used by
the “saveToMatFiles” script. Changing documentation and location of the
files RealNeuralNetAdEx depends on (away from top level and to a folder
in scriptmenu).