SORN script and compile fix
Changed the SORN script which had an error where the superclass
“threshold” variable was being edited instead of the SORN threshold
variable. Also changed some parameters in the SORN script, though
really it’s unclear what they should be without a specific task to test
their optimum values. Also changed the compile script so that it will
work both for the downloaded distribution of Simbrain and the source
code package.