Fix to compile script and SORN color change
Activation for SORN neurons was changed to 2*(netInput-threshold) to
make it look more like the picture on the website (and to make spikes
stand out better). Accordingly SORN neurons now use SYNAPTIC input with
a Step spike responder instead of WEiGHTED and all references to
activation have been switched to a ternary operator expressing
activation of 1 or 0 based on spiking state. The SORN script was
changed to add in the Step spike responders. Compile.sh was changed
because it accidentally didn’t actually compile properly if
Simbrain.jar was found in ./build/main.