Adding an option to update gui components only once every X iterations
Added an option (which is now the default) whereby the gui components in the network panel are only redrawn every X iterations (200 for now is default). Neurons/synapses are however redrawn in response to changes while the network is not running and also if the network is being updated by a one-time single iteration.